requirements.txt
pdf2image==1.16.0
Pillow==9.3.0
pylibdmtx==0.1.10
